% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{getObjsInPipe}
\alias{getObjsInPipe}
\title{Obtain all of objects in the specific pipeline}
\usage{
getObjsInPipe(pipeName = "pipe")
}
\arguments{
\item{pipeName}{\code{Character} scalar or vector. Pipeline name(s) of objects to be selected.}
}
\value{
\code{List} scalar. A list containing all objects that belongs to the pipe name.
}
\description{
Obtain all of objects in the specific pipeline
}
\examples{
getObjsInPipe(pipe)
}
