% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{ignoreCheck}
\alias{ignoreCheck}
\title{ignore checking input and output file (for developer)}
\usage{
ignoreCheck(check = TRUE)
}
\arguments{
\item{check}{\code{Logical} scalar.
Ignore checking input and output file MD5 value when skipping the step.}
}
\value{
\item{ignoreCheck}{No value will be returned}
}
\description{
ignore checking input and output file (for developer)
}
\examples{
ignoreCheck(FALSE)

}
