% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{initPipeFrame}
\alias{initPipeFrame}
\title{initialize the pipeFrame package}
\usage{
initPipeFrame(defaultJobName, availableGenome,
  defaultCheckAndInstallFunc = NULL, defaultThreads = 2,
  defaultTmpDir = getwd(), defaultRefDir = file.path(getwd(),
  "refdir"), defaultReference = list(test = list(file = "fileName", rc =
  "obj")))
}
\arguments{
\item{defaultJobName}{\code{Character} scalar. The default job name for the pacakge.
When user uses pipeFrame package, "pipeFrame-pipeline" is the default defaultJobName.}

\item{availableGenome}{\code{Character} scalar or vector.
To configure the default valid genome like "hg19", "mm10", etc. to be configured.}

\item{defaultCheckAndInstallFunc}{\code{Function} scalar. The function need to call several \code{\link{runWithFinishCheck}}}

\item{defaultThreads}{\code{Numeric} scalar. The maximun thread a step can use. Default 2}

\item{defaultTmpDir}{\code{Character} scalar. The directory for  intermediate result storage of all steps. Default: Current woring directory.}

\item{defaultRefDir}{\code{Character} scalar. The directory for reference data dependency storage. Default: \code{file.path(getwd(),"refdir")}}

\item{defaultReference}{\code{List} scalar. Reference files list.}
}
\value{
Nothing will be returned.
}
\description{
This function should be used first line in R terminal for general user.
 And it shoul be used in .onLoad() function when developing package.
 It's main function is to configure the basic options for the package.
}
\examples{
initPipeFrame(availableGenome = c("hg19", "hg38", "mm9", "mm10"),
              defaultJobName = paste0("pkgname","-pipeline")
)

}
