% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{loadConfig}
\alias{loadConfig}
\alias{saveConfig}
\alias{configRegName}
\title{load configure from file}
\usage{
loadConfig(configFile)

saveConfig(configFile)

configRegName()
}
\arguments{
\item{configFile}{\code{Character} scalar.
The directory to configuration file.}
}
\value{
\item{loadConfig}{No value will be returned}

\item{saveConfig}{save the configuration into a RDS file}

\item{configRegName}{charactor vector, registered configuration name}
}
\description{
load configure from file
}
\examples{
configRegName()
saveConfig("test.rds")
loadConfig("test.rds")

}
