% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{loadStep}
\alias{loadStep}
\title{load step object from rds file}
\usage{
loadStep(rdsfile, regClass = TRUE)
}
\arguments{
\item{rdsfile}{\code{Character} scalar. The rds file directory for Step  (or its inherit class) Object}

\item{regClass}{\code{Logical} scalar. Register the Class of object to inherit from Step Class.
Default: TRUE. Note: make sure corresponding packages depending on pipeFrame is loaded}
}
\value{
Step (or its inherit class) object
}
\description{
load PipeFrame Step (or its inherit class) object from rds file
}
\examples{
objrds <- system.file(package = "pipeFrame", "extdata","pipeFrame.obj.rds")
obj <- loadStep(objrds)

}
