% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{setGenome}
\alias{setGenome}
\alias{getValidGenome}
\alias{getGenome}
\title{Configure genome for all steps}
\usage{
getValidGenome()

setGenome(genome)

getGenome()
}
\arguments{
\item{genome}{\code{Character} scalar. Valid genome to be configured.}
}
\value{
\item{getValidGenome}{\code{Character} scalar. All valid genome assembly for this pacakge.}

\item{setGenome}{All package and dependency are configured or installed. Nothing will be returned.}

\item{getGenome}{\code{Character} scalar. The genome is configured}
}
\description{
Configure genome for all steps
}
\examples{
getValidGenome()
setGenome("hg19")
getGenome()

}
