% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{setPipeName}
\alias{setPipeName}
\alias{getPipeName}
\title{Configure the pipe name for following steps.}
\usage{
setPipeName(pipeName)

getPipeName(all = FALSE)
}
\arguments{
\item{pipeName}{\code{Character} scalar. Job name for following steps.}
}
\value{
\item{setPipeName}{No value will be returned}

\item{getPipeName}{Set a job name for following steps.}
}
\description{
Configure the pipe name for following steps.
}
\examples{
setPipeName("pipe")
getPipeName()
}
