% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{setRefDir}
\alias{setRefDir}
\alias{getRefDir}
\alias{getRef}
\alias{getRefFiles}
\alias{getRefRc}
\title{Configure max threads number for all steps}
\usage{
setRefDir(refdir, createDir = TRUE)

getRefDir()

getRef(refName)

getRefFiles(refName)

getRefRc(refName)
}
\arguments{
\item{refdir}{\code{Character} scalar. The directory for reference data dependency storage.}

\item{createDir}{\code{Logica} scalar. Create the directory when the directory is not exist. Default: TRUE}

\item{refName}{\code{Character} scalar. The reference name.}
}
\value{
\item{setRefDir}{nothing will be returned}

\item{getRefDir}{\code{Character} scalar. The directory for reference data dependency storage.}

\item{getRef}{\code{List} scalar. A list object contain "files" (reference file paths) and "rc" (any R object reference) }

\item{getRefFiles}{\code{Character} scalar or vector.Reference file directory. }

\item{getRefFiles}{\code{Uncertain} scalar or vector.Reference R object. }
}
\description{
Configure max threads number for all steps
}
\examples{
setRefDir("./refdir")
getRefDir()
getRef("test")

getRefFiles("test")
getRefRc("test")

}
