% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{setThreads}
\alias{setThreads}
\alias{getThreads}
\title{Configure the maximum number of threads}
\usage{
setThreads(threads = detectCores())

getThreads()
}
\arguments{
\item{threads}{\code{Numeric} scalar. The maximum number of threads that can be allocated  to each step.}
}
\value{
\item{setThreads}{No value will be returned}

\item{getThreads}{\code{Numeric} scalar. Display the maximum number of threads that can be allocated  to each step.}
}
\description{
Configure the maximum number of threads for all steps
}
\examples{
setThreads()
getThreads()
}
