% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{setTmpDir}
\alias{setTmpDir}
\alias{getTmpDir}
\title{Configure directory for  intermediate result storage of all steps}
\usage{
setTmpDir(tmpDir = getwd())

getTmpDir()
}
\arguments{
\item{tmpDir}{\code{Character} scalar. The directory for  intermediate result storage of all steps. Default: Current woring directory.}
}
\value{
\item{setTmpDir}{nothing will be returned}

\item{getRefDir}{\code{Character} scalar. The directory for  intermediate result storage of all steps.}
}
\description{
Configure directory for  intermediate result storage of all steps
}
\examples{
setTmpDir()
getTmpDir()
}
