% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepBase.R, R/Utils.R
\docType{class}
\name{Step-class}
\alias{Step-class}
\alias{Step}
\alias{init}
\alias{getStepName,Step-method}
\alias{getStepName}
\alias{getDefName,Step-method}
\alias{getDefName}
\alias{getParam,Step-method}
\alias{getParam}
\alias{getParamItems,Step-method}
\alias{getParamItems}
\alias{isReady,Step-method}
\alias{isReady}
\alias{clearProcCache,Step-method}
\alias{clearProcCache}
\alias{getReportVal,Step-method}
\alias{getReportVal}
\alias{getReportItems,Step-method}
\alias{getReportItems}
\alias{getAutoPath,Step-method}
\alias{getAutoPath}
\alias{checkRequireParam}
\alias{checkAllPath}
\alias{getParamMD5Path,Step-method}
\alias{getParamMD5Path}
\alias{getStepWorkDir,Step-method}
\alias{getStepWorkDir}
\alias{getStepId,Step-method}
\alias{getStepId}
\alias{writeLog,Step-method}
\alias{writeLog}
\alias{processing}
\alias{getReportValImp}
\alias{getReportItemsImp}
\alias{checkFileCreatable}
\title{Methods for Step objects}
\usage{
init(.Object, prevSteps = list(), ...)

\S4method{getStepName}{Step}(.Object, ...)

\S4method{getDefName}{Step}(.Object, ...)

\S4method{getParam}{Step}(.Object, item, type = c("input", "output",
  "other"), ...)

\S4method{getParamItems}{Step}(.Object, type = c("input", "output",
  "other"), ...)

\S4method{isReady}{Step}(.Object, ...)

\S4method{clearProcCache}{Step}(.Object, ...)

\S4method{getReportVal}{Step}(.Object, item, ...)

\S4method{getReportItems}{Step}(.Object, ...)

\S4method{getAutoPath}{Step}(.Object, originPath, regexProcName, suffix,
  ...)

checkRequireParam(.Object, ...)

checkAllPath(.Object, ...)

\S4method{getParamMD5Path}{Step}(.Object, ...)

\S4method{getStepWorkDir}{Step}(.Object, ...)

\S4method{getStepId}{Step}(.Object, ...)

\S4method{writeLog}{Step}(.Object, msg, ..., isWarnning = FALSE,
  appendLog = TRUE, showMsg = TRUE)

processing(.Object, ...)

getReportValImp(.Object, item, ...)

getReportItemsImp(.Object, item, ...)

checkFileCreatable(filePath, ...)
}
\arguments{
\item{.Object}{Step object scalar.
Step object return by process functions.}

\item{...}{Additional arguments, currently unused.}

\item{item}{\code{Characters} scalar.
The items in parameter list or report list.}

\item{preProc}{\code{Logitcal} scalar.
show the available upstream processes if TRUE.(Default: FALSE)}

\item{nextProc}{\code{Logitcal} scalar.
show the available downstream processes if TRUE. (Default: TRUE)}

\item{curProc}{\code{Logitcal} scalar.
show the current process of parameter \code{.Object} if TRUE. (Default: TRUE)}

\item{display}{\code{Logitcal} scalar.
Save to pdf file if FALSE. (Default: TRUE)}
}
\value{
the function and result of functions:

\item{init}{(For package developer) An Step child class object with initialized input, output and other parameters}

\item{getStepName}{get Step object Characher name}

\item{getDefName}{get Step object Characher name}

\item{getParam}{Get parameter value set by process function.
See \code{getParamItems} to obtain valid item for query.}

\item{getParamItems}{Get parameter name list}

\item{isReady}{Is the process ready for downstream process}

\item{clearProcCache}{Clear cache of Step object}

\item{getReportVal}{Get report value of item.
See \code{getReportItems} to obtain valid item for query.}

\item{getReportItems}{Get all items that can be reported}

\item{getAutoPath}{(For package developer)Developer can use this method to generate new file name based on exist input file name}

\item{checkRequireParam}{(For package developer) Check required inputs are filled.}

\item{checkRequireParam}{(For package developer) Check required inputs are filled.}

\item{getParamMD5Path}{The Step object storage directory}

\item{getStepWorkDir}{Get the step work directory of this object}

\item{getStepId}{Get the step ID}

\item{writeLog}{(For package developer) write log.}

\item{processing}{(For package developer) Run pipeline step}

\item{getReportValImp}{(For package developer) get Report Value}

\item{getReportItemsImp}{(For package developer) getReportItemsImp}

\item{checkFileCreatable}{(For package developer) Check file creatable.}
}
\description{
You can call Step objects operation methods below to
obtain information in objects.
}
\details{
\code{Step} is a S4 class for generating Step S4 objects.
All Step objects generated by its subclasses.
To generate new Step objects, a function wrapper with fixed auguments needs to
be implemented. User uses these function to generate new Step functions
rather than use Step S4 class to generate object yourself.
}
\examples{

}
\seealso{
\code{\link{setGenome}}
\code{\link{setThread}}
}
\author{
Zheng Wei
}
