% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graphMng}
\alias{graphMng}
\alias{addEdges}
\alias{getPrevSteps}
\alias{getNextSteps}
\alias{printMap}
\title{Step graph management}
\usage{
addEdges(edges, argOrder)

getPrevSteps(stepName, argOrder)

getNextSteps(stepName, argOrder)

printMap(stepName = NULL, display = TRUE, ...)
}
\arguments{
\item{edges}{\code{Character} vector. Contain the start and end points name of all edges.
It need to follow the format like c("startpt1","endpt1","startpt2","endpt2","startpt3","endpt3")}

\item{argOrder}{\code{Numeric} scalar. The augument order of the input Step object.}

\item{stepName}{\code{Character} scalar. Step class name of each step.}
}
\value{
\item{addEdges}{Nother will be returned.}

\item{getPrevSteps}{Previous steps name}

\item{getNextSteps}{Next steps name}

\item{printMap}{print pipeline flow map}
}
\description{
Step graph management
}
