% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{setGenome}
\alias{setGenome}
\alias{getValidGenome}
\alias{getGenome}
\title{Configure genome for all steps}
\usage{
getValidGenome()

setGenome(genome)

getGenome()
}
\arguments{
\item{genome}{\code{Character} scalar. Valid genome to be configured.}
}
\value{
\item{getValidGenome}{\code{Character} scalar. . All valid genome assemblies for this package.}

\item{setGenome}{All packages and dependencies are configured and installed. No value will be returned.}

\item{getGenome}{\code{Character} scalar. Display the configured genome.}
}
\description{
Configure the reference genome assembly for all steps.
}
\examples{
getValidGenome()
setGenome("hg19")
getGenome()

}
