% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{setThreads}
\alias{setThreads}
\alias{getThreads}
\title{Configure max threads number for all steps}
\usage{
setThreads(threads = detectCores())

getThreads()
}
\arguments{
\item{threads}{\code{Numeric} scalar. The max number of threads that can be used by each step.}
}
\value{
\item{setThreads}{nothing will be returned}

\item{getThreads}{\code{Numeric} scalar. The max number of threads that can be used by each step}
}
\description{
Configure max threads number for all steps
}
\examples{
setThreads()
getThreads()
}
