% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configVal.R
\name{setTmpDir}
\alias{setTmpDir}
\alias{getTmpDir}
\title{Configure the directory for intermediate results of all steps}
\usage{
setTmpDir(tmpDir = getwd())

getTmpDir()
}
\arguments{
\item{tmpDir}{\code{Character} scalar.
The directory to store intermediate results of all steps.
Default: Current directory.}
}
\value{
\item{setTmpDir}{No value will be returned}

\item{getTmpDir}{\code{Character} scalar.
Display the directory for intermediate results of all steps.}
}
\description{
Configure the directory for intermediate results of all steps
}
\examples{
setTmpDir()
getTmpDir()
}
