% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils}
\alias{Utils}
\alias{getBasenamePrefix}
\alias{getPathPrefix}
\alias{checkFileExist}
\alias{checkPathExist}
\alias{checkFileCreatable}
\title{functions for directory operations}
\usage{
getBasenamePrefix(filepath, words, ...)

getPathPrefix(filepath, words, ...)

checkFileExist(filePath, ...)

checkPathExist(filePath, ...)

checkFileCreatable(filePath, ...)
}
\arguments{
\item{filepath}{\code{character} scalar or vector.}

\item{words}{\code{character} scalar}
}
\value{
\item{getBasenamePrefix}{Get the step ID}

\item{getPathPrefix}{Get the step ID}

\item{checkFileExist}{(For package developer) Check file is exist.}

\item{checkPathExist}{(For package developer) Check directory is exist.}

\item{checkFileCreatable}{(For package developer) Check file creatable.}
}
\description{
functions for directory operations
}
