% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils}
\alias{Utils}
\alias{getBasenamePrefix}
\alias{getPathPrefix}
\alias{checkFileExist}
\alias{checkPathExist}
\alias{checkFileCreatable}
\title{functions for directory operations}
\usage{
getBasenamePrefix(filepath, words, ...)

getPathPrefix(filepath, words, ...)

checkFileExist(filePath, ...)

checkPathExist(filePath, ...)

checkFileCreatable(filePath, ...)
}
\arguments{
\item{filepath}{\code{character} scalar or vector.}

\item{words}{\code{character} scalar. Remove substring of the path characters starting to match the word}

\item{...}{Additional arguments, currenly unused}

\item{filePath}{\code{Character} scalar.}
}
\value{
\item{getBasenamePrefix}{Get the filepath  basename with removed suffix}

\item{getPathPrefix}{Get the filepath  with removed suffix}

\item{checkFileExist}{(For package developer) Check file is exist.}

\item{checkPathExist}{(For package developer) Check directory is exist.}

\item{checkFileCreatable}{(For package developer) Check file creatable.}
}
\description{
functions for directory operations
}
\examples{
getBasenamePrefix("aaa/bbb.ccc.ddd","cCc")

getBasenamePrefix("aaa/bbb.ccc.ddd","ddd")

getPathPrefix("aaa/bbb.ccc.ddd","dDd")

getPathPrefix("aaa/bbb.ccc.ddd","ccc")

file.create("test.bed")

checkFileExist("test.bed")

tryCatch({checkFileExist("test.bed1")},error = function(e) e)

dir.create("testdir")

checkPathExist("testdir")

tryCatch({checkPathExist("testdir1")},error = function(e) e)

checkFileCreatable("aaa.bed")

tryCatch({checkFileCreatable("testdir1/aaa.bed")},error = function(e) e)


}
